+++ r/tests/testthat/test-pseudobulk.R
@@ -29,7 +29,7 @@ test_that("Pseudobulk aggregation works", {
   m0 <- as.matrix(m0)
   m0_t <- t(m0)
   m1_t <- t(m1)
-  for (matrices in list(list(m0, m1), list(m0_t, m1_t))) {
+  for (matrices in list(list(m0_t, m1_t), list(m0, m1))) {
     # Check across two equal groups, one group, and groups of equal length
     m <- matrices[[1]]
     m_bpcells <- matrices[[2]]
@@ -37,34 +37,67 @@ test_that("Pseudobulk aggregation works", {
     groups_one_type <- c(rep.int(1, ncol(m)))
     groups_equal_length <- seq(ncol(m))
     for (cell_group in list(groups, groups_one_type, groups_equal_length)) {
-      # create clipping
-      min_vals  <- m %>% apply(2, quantile, 0.99, type = 1)
-      #pmin doesn't work correctly when comparing matrix to vec, so we do it iteratively
       m_sum <- create_pseudobulk_r(m, cell_group, "sum")
       m_mean <- create_pseudobulk_r(m, cell_group, "mean")
       m_non_zeros <- create_pseudobulk_r(m, cell_group, "non_zeros")
-      m_bpcells_sum <- pseudobulk_matrix(m_bpcells, cell_group, method = "sum")$sum
-      m_bpcells_mean <- pseudobulk_matrix(m_bpcells, cell_group, method = "mean")$mean
-      m_bpcells_non_zeros <- pseudobulk_matrix(m_bpcells, cell_group, method = "non-zeros")$non_zeros
+      m_bpcells_sum <- pseudobulk_matrix(m_bpcells, cell_group, method = "sum")
+      m_bpcells_mean <- pseudobulk_matrix(m_bpcells, cell_group, method = "mean")
+      m_bpcells_non_zeros <- pseudobulk_matrix(m_bpcells, cell_group, method = "nonzeros")
       expect_equal(m_sum, m_bpcells_sum)
       # check for clipping if matrix isn't transposed
       if (m_bpcells@transpose == FALSE) {
-        m_clipped <- m
-        m_sum_clipped <- create_pseudobulk_r(m, cell_group, "sum")
-        m_bpcells_clipped <- pseudobulk_matrix(m_bpcells, cell_group, method = "sum", clip_values = TRUE)$sum
-        expect_equal(m_sum_clipped, m_bpcells_clipped)
+        for (quantile in c(0.1, 0.25, 0.5, 0.75, 0.99)) {
+          min_vals  <- m %>% apply(2, quantile, probs = quantile, type = 7) %>% unlist()
+          m_clipped <- t(pmin(t(m), min_vals))
+          m_sum_clipped <- create_pseudobulk_r(m_clipped, cell_group, "sum")
+          m_bpcells_clipped <- pseudobulk_matrix(m_bpcells, cell_group, method = "sum", clip_values = quantile)
+          expect_equal(m_sum_clipped, m_bpcells_clipped)
+        }
       }
       expect_equal(m_mean, m_bpcells_mean)
       expect_equal(m_non_zeros, m_bpcells_non_zeros)
       # make sure that we dont check for variances if we have number of groups == number of cells
       if (length(unique(cell_group)) < ncol(m)) {
-        m_clipped <- m
-        for (idx in seq_along(min_vals)) {
-          m_clipped[, idx] <- pmin(m[, idx], min_vals[[idx]])
-        }
         m_var <- create_pseudobulk_r(m, cell_group, "var")
         m_bpcells_var <- pseudobulk_matrix(m_bpcells, cell_group, method = "var")
-        expect_equal(m_var, m_bpcells_var$var)
+        expect_equal(m_var, m_bpcells_var)
+      }
+    }
+  }
+})
+
+test_that("Pseudobulk aggregation works with multiple return types", {
+  m0 <- generate_sparse_matrix(20, 10, max_val = 10)
+  m1 <- m0 |> as("IterableMatrix")
+  m0 <- as.matrix(m0)
+  m0_t <- t(m0)
+  m1_t <- t(m1)
+  methods <- c("nonzeros", "sum", "mean", "var")
+  for (matrices in list(list(m0_t, m1_t), list(m0, m1))) {
+    # Check across two equal groups, one group, and groups of equal length
+    m <- matrices[[1]]
+    m_bpcells <- matrices[[2]]
+    cell_group <- c(rep.int(1, ncol(m)/2), rep.int(2, ncol(m)/2))
+    m_sum <- create_pseudobulk_r(m, cell_group, "sum")
+    m_mean <- create_pseudobulk_r(m, cell_group, "mean")
+    m_non_zeros <- create_pseudobulk_r(m, cell_group, "non_zeros")
+    m_var <- create_pseudobulk_r(m, cell_group, "var")
+    for (first_method_idx in seq(1, length(methods)-1)) {
+      for (second_method_idx in seq(first_method_idx+1, length(methods))) {
+        methods_of_interest <- c(methods[first_method_idx], methods[second_method_idx])
+        m_bpcells_res <- pseudobulk_matrix(m_bpcells, cell_group, method = methods_of_interest)
+        if ("var" %in% methods_of_interest) {
+          expect_equal(m_var, m_bpcells_res$var)
+        }
+        if ("mean" %in% methods_of_interest) {
+          expect_equal(m_mean, m_bpcells_res$mean)
+        }
+        if ("sum" %in% methods_of_interest) {
+          expect_equal(m_sum, m_bpcells_res$sum)
+        }
+        if ("nonzeros" %in% methods_of_interest) {
+          expect_equal(m_non_zeros, m_bpcells_res$nonzeros)
+        }
       }
     }
   }
@@ -75,8 +108,8 @@ test_that("Matrix quantiles work", {
   m1 <- m0 |> as("IterableMatrix")
   m0 <- as.matrix(m0)
   for (quantile in c(0, 0.25, 0.5, 0.75, 0.99)) {
-    m_quantile <- tibble::as_tibble(m0) %>% apply(2, quantile, probs = quantile, type = 1)
-    m_quantile_bpcells <- matrix_quantile_per_cell(m1, quantile)
+    m_quantile <- tibble::as_tibble(m0) %>% apply(2, quantile, probs = quantile, type = 7)
+    m_quantile_bpcells <- colQuantiles(m1, quantile)
     expect_equal(m_quantile, m_quantile_bpcells)
   }
 })