+++ r/NEWS.md
@@ -19,6 +19,9 @@ Contributions welcome :)
   Thanks to @ycli1995 for pull request #110
 - Add `trackplot_genome_annotation()` for plotting peaks, with options for directional arrows, colors, labels, and peak widths. (pull request #113)
 - Add MACS2/3 input creation and peak calling through `call_macs_peaks()` (pull request #118)
+- Add `colQuantiles()` which returns the quantiles of each column of a matrix. 
+  If `matrixStats` or `MatrixGenerics` packages are installed, `BPCells::colQuantiles()` will fall back to their implementations for non-BPCells objects. (pull request #128)
+- Add `pseudobulk_matrix()` which allows for creation of a pseudobulk matrix aggregated by `sum`, `nonzeros`, `mean`, `var`. (pull request #128)
 
 ## Improvements
 - `trackplot_loop()` now accepts discrete color scales